+++ simplebluez/src/BluezRoot.cpp
@@ -5,9 +5,10 @@
 using namespace SimpleBluez;
 
 BluezRoot::BluezRoot(std::shared_ptr<SimpleDBus::Connection> conn, const std::string& bus_name, const std::string& path)
-    : Proxy(conn, bus_name, path) {
+    : Proxy(conn, bus_name, path) {}
 
-    _interfaces.emplace(std::make_pair("org.freedesktop.DBus.ObjectManager", std::make_shared<SimpleDBus::ObjectManager>(conn, shared_from_this())));
+void BluezRoot::on_registration() {
+    _interfaces.emplace(std::make_pair("org.freedesktop.DBus.ObjectManager", std::make_shared<SimpleDBus::ObjectManager>(_conn, shared_from_this())));
 
     object_manager()->InterfacesAdded = [&](std::string path, SimpleDBus::Holder options) { path_add(path, options); };
     object_manager()->InterfacesRemoved = [&](std::string path, SimpleDBus::Holder options) {