+++ docs/changelog.rst
@@ -19,6 +19,7 @@ The format is based on `Keep a Changelog`_, and this project adheres to `Semanti
 - Configuration option to select which BlueZ backend to use.
 - (Android) Added support to set and retrieve the JavaVM pointer.
 - (Linux) Added frozen BlueZ backend in preparation for upcoming changes.
+- (SimpleDBus) Added dedicated Properties interface.
 
 **Changed**